Add: input regression callback14

+++ src/omv/modules/py_tf.c
@@ -275,7 +275,7 @@ STATIC void py_tf_regression_input_callback(void *callback_data,
     if(params->input_channels == 1) {
         if (params->input_datatype == LIBTF_DATATYPE_INT8) {
 
-            float *model_input_i64 = (float *) model_input;
+            int64_t *model_input_i64 = (int64_t *) model_input;
             
             for(; size >= 0; size -=1){
                 // float temp = (float)(arg->input_list->items[size]);
